+++ CHANGES.md
@@ -1,1 +1,7 @@
 # Changes Logs
+
+## v0.2.0
+
+- Added: method `findExpiringCertificates` for class `CertificateManager`.
+- Added: method `clear` for class `CertificateManager`.
+- Fixed: method `use` of class `CertificateManager` can not update certificate.